# gin-front-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "博客用户后台API接口文档",
        "title": "博客用户后台",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/bArticle/add/comment": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "新增文章评论",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章评论相关接口"
                ],
                "summary": "新增文章评论",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AddCommentDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/result.Result"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.AddCommentDto": {
            "type": "object",
            "required": [
                "articleId",
                "content"
            ],
            "properties": {
                "articleId": {
                    "description": "文章ID",
                    "type": "integer"
                },
                "content": {
                    "description": "回复内容",
                    "type": "string"
                },
                "pCommentId": {
                    "description": "父级评论ID",
                    "type": "integer"
                },
                "replyUserId": {
                    "description": "回复人的id",
                    "type": "integer"
                }
            }
        },
...
        "result.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "返回的数据"
                },
                "message": {
                    "description": "提示信息",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}